refactor: Move TestResult classes to probes
