+++ src/main/java/de/rub/nds/scanner/core/guideline/GuidelineCheckCondition.java
@@ -9,8 +9,8 @@
 package de.rub.nds.scanner.core.guideline;
 
 import de.rub.nds.scanner.core.constants.AnalyzedProperty;
-import de.rub.nds.scanner.core.constants.TestResult;
-import de.rub.nds.scanner.core.constants.TestResults;
+import de.rub.nds.scanner.core.probe.result.TestResult;
+import de.rub.nds.scanner.core.probe.result.TestResults;
 import jakarta.xml.bind.annotation.*;
 import java.util.List;
