+++ src/main/java/de/rub/nds/scanner/core/guideline/GuidelineChecker.java
@@ -8,7 +8,7 @@
  */
 package de.rub.nds.scanner.core.guideline;
 
-import de.rub.nds.scanner.core.constants.TestResults;
+import de.rub.nds.scanner.core.probe.result.TestResults;
 import de.rub.nds.scanner.core.report.ScanReport;
 import java.util.ArrayList;
 import java.util.List;
