+++ src/main/java/de/rub/nds/scanner/core/probe/result/CollectionResult.java
@@ -6,7 +6,7 @@
  * Licensed under Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0.txt
  */
-package de.rub.nds.scanner.core.constants;
+package de.rub.nds.scanner.core.probe.result;
 
 import java.io.Serializable;
 import java.util.Collection;
