+++ src/main/java/de/rub/nds/scanner/core/report/rating/RatingInfluencer.java
@@ -9,7 +9,7 @@
 package de.rub.nds.scanner.core.report.rating;
 
 import de.rub.nds.scanner.core.constants.AnalyzedProperty;
-import de.rub.nds.scanner.core.constants.TestResult;
+import de.rub.nds.scanner.core.probe.result.TestResult;
 import jakarta.xml.bind.annotation.*;
 import java.util.Arrays;
 import java.util.LinkedList;
